\name{print.saws}
\alias{print.saws}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Print saws object}
\description{Prints confidence intervals and p-values from saws object.}
\usage{
\method{print}{saws}(x, digits = NULL, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{object of class 'saws'}
  \item{digits}{ number of digits}
  \item{\dots}{ other objects passed to print default}
}
%\details{}
%\examples{}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{misc}
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
